+++ excel/auto/hero_entry.go
@@ -11,15 +11,15 @@ var	heroEntries	*HeroEntries	//hero.xlsx全局变量
 
 // hero.xlsx属性表
 type HeroEntry struct {
-	Id        	int       	`json:"Id,omitempty"`	//id        
+	Id        	int32     	`json:"Id,omitempty"`	//id        
 	Name      	string    	`json:"Name,omitempty"`	//名字        
-	AttID     	int       	`json:"AttID,omitempty"`	//属性id      
-	Quality   	int       	`json:"Quality,omitempty"`	//品质        
+	AttID     	int32     	`json:"AttID,omitempty"`	//属性id      
+	Quality   	int32     	`json:"Quality,omitempty"`	//品质        
 }
 
 // hero.xlsx属性表集合
 type HeroEntries struct {
-	Rows      	map[int]*HeroEntry	`json:"Rows,omitempty"`	//          
+	Rows      	map[int32]*HeroEntry	`json:"Rows,omitempty"`	//          
 }
 
 func  init()  {
@@ -29,7 +29,7 @@ func  init()  {
 func (e *HeroEntries) Load(excelFileRaw *excel.ExcelFileRaw) error {
 	
 	heroEntries = &HeroEntries{
-		Rows: make(map[int]*HeroEntry),
+		Rows: make(map[int32]*HeroEntry),
 	}
 
 	for _, v := range excelFileRaw.CellData {
@@ -48,12 +48,12 @@ func (e *HeroEntries) Load(excelFileRaw *excel.ExcelFileRaw) error {
 	
 }
 
-func  GetHeroEntry(id int) (*HeroEntry, bool) {
+func  GetHeroEntry(id int32) (*HeroEntry, bool) {
 	entry, ok := heroEntries.Rows[id]
 	return entry, ok
 }
 
-func  GetHeroSize() int {
-	return len(heroEntries.Rows)
+func  GetHeroSize() int32 {
+	return int32(len(heroEntries.Rows))
 }
