+++ excel/auto/playerLevelup_entry.go
@@ -11,13 +11,13 @@ var	playerLevelupEntries	*PlayerLevelupEntries	//playerLevelup.xlsx全局变量
 
 // playerLevelup.xlsx属性表
 type PlayerLevelupEntry struct {
-	Id        	int       	`json:"Id,omitempty"`	//等级        
-	Exp       	int       	`json:"Exp,omitempty"`	//达到此等级需要的经验值
+	Id        	int32     	`json:"Id,omitempty"`	//等级        
+	Exp       	int32     	`json:"Exp,omitempty"`	//达到此等级需要的经验值
 }
 
 // playerLevelup.xlsx属性表集合
 type PlayerLevelupEntries struct {
-	Rows      	map[int]*PlayerLevelupEntry	`json:"Rows,omitempty"`	//          
+	Rows      	map[int32]*PlayerLevelupEntry	`json:"Rows,omitempty"`	//          
 }
 
 func  init()  {
@@ -27,7 +27,7 @@ func  init()  {
 func (e *PlayerLevelupEntries) Load(excelFileRaw *excel.ExcelFileRaw) error {
 	
 	playerLevelupEntries = &PlayerLevelupEntries{
-		Rows: make(map[int]*PlayerLevelupEntry),
+		Rows: make(map[int32]*PlayerLevelupEntry),
 	}
 
 	for _, v := range excelFileRaw.CellData {
@@ -46,12 +46,12 @@ func (e *PlayerLevelupEntries) Load(excelFileRaw *excel.ExcelFileRaw) error {
 	
 }
 
-func  GetPlayerLevelupEntry(id int) (*PlayerLevelupEntry, bool) {
+func  GetPlayerLevelupEntry(id int32) (*PlayerLevelupEntry, bool) {
 	entry, ok := playerLevelupEntries.Rows[id]
 	return entry, ok
 }
 
-func  GetPlayerLevelupSize() int {
-	return len(playerLevelupEntries.Rows)
+func  GetPlayerLevelupSize() int32 {
+	return int32(len(playerLevelupEntries.Rows))
 }
