+++ excel/auto/runeSuit_entry.go
@@ -11,17 +11,17 @@ var	runeSuitEntries	*RuneSuitEntries	//runeSuit.xlsx全局变量
 
 // runeSuit.xlsx属性表
 type RuneSuitEntry struct {
-	Id        	int       	`json:"Id,omitempty"`	//id        
-	Suit2_AttID	int       	`json:"Suit2_AttID,omitempty"`	//2件套装属性ID  
-	Suit3_AttID	int       	`json:"Suit3_AttID,omitempty"`	//3件套装属性ID  
-	Suit4_AttID	int       	`json:"Suit4_AttID,omitempty"`	//2件套装属性ID  
-	Suit5_AttID	int       	`json:"Suit5_AttID,omitempty"`	//2件套装属性ID  
-	Suit6_AttID	int       	`json:"Suit6_AttID,omitempty"`	//2件套装属性ID  
+	Id        	int32     	`json:"Id,omitempty"`	//id        
+	Suit2_AttID	int32     	`json:"Suit2_AttID,omitempty"`	//2件套装属性ID  
+	Suit3_AttID	int32     	`json:"Suit3_AttID,omitempty"`	//3件套装属性ID  
+	Suit4_AttID	int32     	`json:"Suit4_AttID,omitempty"`	//2件套装属性ID  
+	Suit5_AttID	int32     	`json:"Suit5_AttID,omitempty"`	//2件套装属性ID  
+	Suit6_AttID	int32     	`json:"Suit6_AttID,omitempty"`	//2件套装属性ID  
 }
 
 // runeSuit.xlsx属性表集合
 type RuneSuitEntries struct {
-	Rows      	map[int]*RuneSuitEntry	`json:"Rows,omitempty"`	//          
+	Rows      	map[int32]*RuneSuitEntry	`json:"Rows,omitempty"`	//          
 }
 
 func  init()  {
@@ -31,7 +31,7 @@ func  init()  {
 func (e *RuneSuitEntries) Load(excelFileRaw *excel.ExcelFileRaw) error {
 	
 	runeSuitEntries = &RuneSuitEntries{
-		Rows: make(map[int]*RuneSuitEntry),
+		Rows: make(map[int32]*RuneSuitEntry),
 	}
 
 	for _, v := range excelFileRaw.CellData {
@@ -50,12 +50,12 @@ func (e *RuneSuitEntries) Load(excelFileRaw *excel.ExcelFileRaw) error {
 	
 }
 
-func  GetRuneSuitEntry(id int) (*RuneSuitEntry, bool) {
+func  GetRuneSuitEntry(id int32) (*RuneSuitEntry, bool) {
 	entry, ok := runeSuitEntries.Rows[id]
 	return entry, ok
 }
 
-func  GetRuneSuitSize() int {
-	return len(runeSuitEntries.Rows)
+func  GetRuneSuitSize() int32 {
+	return int32(len(runeSuitEntries.Rows))
 }
