+++ excel/auto/skill_entry.go
@@ -11,26 +11,26 @@ var	skillEntries	*SkillEntries	//skill.xlsx全局变量
 
 // skill.xlsx属性表
 type SkillEntry struct {
-	Id        	int       	`json:"Id,omitempty"`	//id        
+	Id        	int32     	`json:"Id,omitempty"`	//id        
 	Name      	string    	`json:"Name,omitempty"`	//名字        
 	Desc      	string    	`json:"Desc,omitempty"`	//描述        
 	Icon      	string    	`json:"Icon,omitempty"`	//图标        
-	NextLevel 	int       	`json:"NextLevel,omitempty"`	//下个等级      
+	NextLevel 	int32     	`json:"NextLevel,omitempty"`	//下个等级      
 	CD        	float32   	`json:"CD,omitempty"`	//技能CD      
-	SkillCombo	int       	`json:"SkillCombo,omitempty"`	//连击类型      
-	SkillType 	int       	`json:"SkillType,omitempty"`	//技能类型      
-	Precondition	int       	`json:"Precondition,omitempty"`	//前置条件      
-	TargetType	int       	`json:"TargetType,omitempty"`	//释放目标类型    
-	TargetRule	[]int     	`json:"TargetRule,omitempty"`	//释放目标规则    
-	AttackType	int       	`json:"AttackType,omitempty"`	//攻击方式      
-	SkillBlocks	[]int     	`json:"SkillBlocks,omitempty"`	//包含的技能块    
-	EffectType	int       	`json:"EffectType,omitempty"`	//特效类型      
+	SkillCombo	int32     	`json:"SkillCombo,omitempty"`	//连击类型      
+	SkillType 	int32     	`json:"SkillType,omitempty"`	//技能类型      
+	Precondition	int32     	`json:"Precondition,omitempty"`	//前置条件      
+	TargetType	int32     	`json:"TargetType,omitempty"`	//释放目标类型    
+	TargetRule	[]int32   	`json:"TargetRule,omitempty"`	//释放目标规则    
+	AttackType	int32     	`json:"AttackType,omitempty"`	//攻击方式      
+	SkillBlocks	[]int32   	`json:"SkillBlocks,omitempty"`	//包含的技能块    
+	EffectType	int32     	`json:"EffectType,omitempty"`	//特效类型      
 	Resource  	string    	`json:"Resource,omitempty"`	//特效路径      
 }
 
 // skill.xlsx属性表集合
 type SkillEntries struct {
-	Rows      	map[int]*SkillEntry	`json:"Rows,omitempty"`	//          
+	Rows      	map[int32]*SkillEntry	`json:"Rows,omitempty"`	//          
 }
 
 func  init()  {
@@ -40,7 +40,7 @@ func  init()  {
 func (e *SkillEntries) Load(excelFileRaw *excel.ExcelFileRaw) error {
 	
 	skillEntries = &SkillEntries{
-		Rows: make(map[int]*SkillEntry),
+		Rows: make(map[int32]*SkillEntry),
 	}
 
 	for _, v := range excelFileRaw.CellData {
@@ -59,12 +59,12 @@ func (e *SkillEntries) Load(excelFileRaw *excel.ExcelFileRaw) error {
 	
 }
 
-func  GetSkillEntry(id int) (*SkillEntry, bool) {
+func  GetSkillEntry(id int32) (*SkillEntry, bool) {
 	entry, ok := skillEntries.Rows[id]
 	return entry, ok
 }
 
-func  GetSkillSize() int {
-	return len(skillEntries.Rows)
+func  GetSkillSize() int32 {
+	return int32(len(skillEntries.Rows))
 }
